Adds move cuisineportal to be under apps
